Move ContentTable component into it's parent's file
